fix int to double

+++ app/src/main/java/ru/vlasov/carmanager/models/Car.kt
@@ -2,9 +2,9 @@ package ru.vlasov.carmanager.models
 
 
 data class Car (val id : Long = 0, val name : String,
-    val yearOfIssue: String, val mileageInKm : Int,
+    val yearOfIssue: String, val mileageInKm : Double,
     val automobileBody : String, val color : String,
-    val engine : CarEngine, val taxPerYear : Int,
+    val engine : CarEngine, val taxPerYear : Double,
     val transmissionType : String, val typeOfDriveUnit : String,
     val steeringWheelLocation : String, val VIN : String,
     val stateNumber : String, val description : String)
